Add a basic homescreen and navigator.
